refactor: Rename spark/borrow to spark/auto
